Migrate InsertPlanV to hawk_main::insert submodule
